Move `parse_ref()` tests to `refs/mod.rs`
